Project: resolve_interpreter(): add in_import parameter
When in_import=True, resolve_interpreter() will not memorize the result
of the resolution process, and will not create any virtual environments.

This flag will be set when resolve_interpreter() is called from within
the import path, when the project model is not yet initialized
and the true constraint (requires-python) is not yet known.