Project: introduce `Project.is_not_local`
Add a new `Project.is_not_local` property that is a "stronger"
variant of `Project.is_global`; that is, it reflects whether a
`pyproject.toml` was actually found, or we are using some kind
of a substitute (either the global project if `pdm --global`
or if `global_project.fallback` is set, or $PWD if neither).

This property would be used to determine if we have to create
a virtualenv in `$PWD` (`$PWD/.venv`) or in `$XDG_DATA_HOME`.