commands: venv: try better to determine the virtualenv prompt
Always use the virtualenv name (`pdm venv create -n`) for the prompt
if passed, otherwise use the last component of the target location
(except `.venv` if the virtualenv is created in-project).

This way, if one creates a venv with `pdm venv create -n myvenv`
or `pdm venv create -p /path/to/myvenv`, we end up with a `(myvenv-X.Y)`
prompt rather than `(<project>-X.Y)` (or even `(<bogusname>-X.Y)` if the
project does not actually exist and we are using the global project or
the `$PWD` stub).