exceptions: strip markup from rich-text warnings for `.__str__()`
Strip markup from the warning text before passing it to the Warning
constructor. The original full warning text will be available via
`warning.full()` instead of `str(warning)`. This way, we do not pollute
the output with unrendered markup if a warning happens to be handled
by the default Python warning system.