pdm, termui: emit spinners on stderr
Given that all high-level log messages are emitted on stderr
(see termui.UI.*), also emit the spinners on stderr to avoid garbled
output when log messages get intermixed with the spinner.

Rich has built-in support for pausing the spinner for text output,
but this support requires that both are output on the same console.
Make use of that.