pdm: write the lock file after the spinner is done
The accompanying message is emitted on stdout, which clashes with the
spinner which is now emitted on stderr (following all other logging).

This is not resolvable without either moving all output to a single
console or teaching python-rich some sort of console interworking.
Both options are way too invasive, so just move the lock file writing
outside of the spinner context manager.